change Sample typedef to be float instead of double - was making bad sounds bro trust me

+++ src/constants.h
@@ -16,4 +16,4 @@
 #define M_PI  (3.14159265)
 #endif
 
-typedef double Sample;
+typedef float Sample;
